% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{windpw}
\alias{windpw}
\title{Wind turbine operational data}
\format{A data frame with 1000 rows and 6 variables:
\itemize{
  \item \code{V}: wind speed (\eqn{m/s}),
  \item \code{D}: wind direction (degree),
  \item \code{rho}: air density (\eqn{kg/m^3}),
  \item \code{I}: turbulence intensity,
  \item \code{Sb}: below-hub wind shear,
  \item \code{y}: normalized power output relative to the rated power (\%).
}}
\usage{
windpw
}
\description{
A dataset containing the measurements of wind-related and other environmental variables as
well as the actual power output measurements of an operating wind turbine.
}
\note{
This dataset is a subset of an actual operational dataset, which is available at
\url{http://ise.tamu.edu/metrology/} where other operational datasets are also available.
To access the datasets, go to `Publications' tab and click the link `data' attached to J53.

This dataset was generated by drawing 1000 random samples from the original dataset.
As such, the sequence of rows is not arranged in time.
}
\keyword{datasets}
